pyproject.toml
[project]
name = "project_KuberAi"
version = "0.1.0"
description = "KuberAI - A multi-component AI project with MCP server and client"
authors = [
    {name = "Shreyas", email = "shreyasbhat17@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
license = {text = "MIT"}

# Common dependencies across all components
dependencies = [
    "kubernetes>=33.1.0",
    "pydantic>=2.0.0",
    "requests>=2.32.4",
    "typing-extensions>=4.0.0",
    "fastmcp>=2.10.1",
    "langchain>=0.3.26",
    "langchain-google-genai>=2.1.6",
    "fastapi>=0.115.14",
    "langgraph>=0.5.1",
    "black>=25.1.0",
    "pre-commit>=4.2.0",
    "ruff>=0.12.1",
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-asyncio>=0.21.0",
    "langchain-mcp-adapters>=0.1.8",
]

# Hatchling Config
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = [
    "mcp_client/src",
    "mcp_server/src",
]

# Tool configurations
[tool.black]
line-length = 120
target-version = ['py312']

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "B", "I", "D"]
fixable = ["ALL"]
unfixable = []

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"

# Workspace configuration for managing multiple components
[tool.uv.workspace]
members = [
    "mcp_server",
    "mcp_client",
]
